Move the LegacyCourse model to a module with the same name
