+++ lms/models/legacy_course.py
